+++ ocpmodels/datasets/__init__.py
@@ -3,7 +3,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .dataset import (
+from .lmdb_dataset import (
     LmdbDataset,
     SinglePointLmdbDataset,
     TrajectoryLmdbDataset,
