+++ ocpmodels/trainers/energy_trainer.py
@@ -41,8 +41,6 @@ class EnergyTrainer(BaseTrainer):
             (default: :obj:`None`)
         is_debug (bool, optional): Run in debug mode.
             (default: :obj:`False`)
-        is_vis (bool, optional): Run in debug mode.
-            (default: :obj:`False`)
         is_hpo (bool, optional): Run hyperparameter optimization with Ray Tune.
             (default: :obj:`False`)
         print_every (int, optional): Frequency of printing logs.
@@ -70,7 +68,6 @@ class EnergyTrainer(BaseTrainer):
         timestamp_id=None,
         run_dir=None,
         is_debug=False,
-        is_vis=False,
         is_hpo=False,
         print_every=100,
         seed=None,
@@ -90,7 +87,6 @@ class EnergyTrainer(BaseTrainer):
             timestamp_id=timestamp_id,
             run_dir=run_dir,
             is_debug=is_debug,
-            is_vis=is_vis,
             is_hpo=is_hpo,
             print_every=print_every,
             seed=seed,
